pyproject.toml
[tool.poetry]
name = "restapiboys"
version = "0.1.0"
description = "A REST API framework Based On YAML Specifications"
authors = ["Ewen Le Bihan <ewen.lebihan7@gmail.com>"]
license = "MIT"

[tool.poetry.scripts]
restapiboys = "restapiboys.cli:run"

[tool.poetry.dependencies]
python = "^3.7"
termcolor = "^1.1.0"
colorama = "^0.4.3"
gunicorn = "19.10.0"
peewee = "^3.13.2"
pyyaml = "^5.3.1"
python-dotenv = "^0.13.0"
docopt = "^0.6.2"
zipp = "1.2.0"
jsonschema = "^3.2.0"

[tool.poetry.dev-dependencies]
black = {version = "^19.10b0", allow-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
